refactor(port): remove deprecated from_chars functions from header
Encourage people to use parse_number_exact instead of from_chars by removing
from_chars from <quick-lint-js/port/integer.h>.
